+++ src/components/chat-details/chat-details.component.ts
@@ -9,8 +9,10 @@ import {
   storeHoc,
   StoreHocProps,
 } from 'hocs';
-import { MessageModel, UserModel } from 'models';
-import { ChatsService, UserService } from 'services';
+import { MessageModel } from 'models';
+import { ChatsService } from 'services';
+import { ModalAddUserComponent } from '../modal-add-user/modal-add-user.component';
+import { ModalChatUsersComponent } from '../modal-chat-users/modal-chat-users.component';
 
 export type ChatDetailsComponentProps = RouterHocProps &
   ActiveChatHocProps &
@@ -19,23 +21,13 @@ export type ChatDetailsComponentProps = RouterHocProps &
     onPopupOpen: () => void;
     onPopupClose: () => void;
     popupIsOpened?: boolean;
-
-    onModalChatUsersOpen: () => void;
-    onModalChatUsersClose: () => void;
-    modalChatUsersIsOpened?: boolean;
-    chatUsers: () => UserModel[] | null;
-
-    onModalAddUserOpen: () => void;
-    onModalAddUserClose: () => void;
-    modalAddUserIsOpened?: boolean;
-
-    searchUsers: () => UserModel[] | null;
-    addUser: (event: MouseEvent) => void;
-    onInputSearchUser: (event: InputEvent) => void;
-    onSearchUser: (event: MouseEvent) => void;
-
     deleteChats: () => void;
-    deleteUser: (event: MouseEvent) => void;
+
+    // Add user
+    onModalAddUserOpen: () => void;
+
+    // Chat users
+    onModalChatUsersOpen: () => void;
   };
 
 export class ChatDetailsComponent extends Block<ChatDetailsComponentProps> {
@@ -49,21 +41,10 @@ export class ChatDetailsComponent extends Block<ChatDetailsComponentProps> {
     this.setProps({
       onPopupOpen: this.onPopupOpen.bind(this),
       onPopupClose: this.onPopupClose.bind(this),
-
       deleteChats: this.deleteChats.bind(this),
-      deleteUser: this.deleteUser.bind(this),
-      addUser: this.addUser.bind(this),
-      onInputSearchUser: this.onInputSearchUser.bind(this),
-      onSearchUser: this.onSearchUser.bind(this),
-
-      onModalChatUsersOpen: this.onModalChatUsersOpen.bind(this),
-      onModalChatUsersClose: this.onModalChatUsersClose.bind(this),
 
       onModalAddUserOpen: this.onModalAddUserOpen.bind(this),
-      onModalAddUserClose: this.onModalAddUserClose.bind(this),
-
-      chatUsers: () => this.props.store.getState().chatUsers,
-      searchUsers: () => this.props.store.getState().searchUsers,
+      onModalChatUsersOpen: this.onModalChatUsersOpen.bind(this),
     });
   }
 
@@ -79,18 +60,6 @@ export class ChatDetailsComponent extends Block<ChatDetailsComponentProps> {
     }
   }
 
-  override _componentWillUnmount() {
-    super._componentWillUnmount();
-  }
-
-  override componentDidMount(props: ChatDetailsComponentProps) {
-    super.componentDidMount(props);
-  }
-
-  override componentWillUnmount() {
-    super.componentWillUnmount();
-  }
-
   onPopupOpen(): void {
     this.setProps({
       popupIsOpened: true,
@@ -103,32 +72,6 @@ export class ChatDetailsComponent extends Block<ChatDetailsComponentProps> {
     });
   }
 
-  onModalChatUsersOpen(): void {
-    this.setProps({
-      modalChatUsersIsOpened: true,
-      popupIsOpened: false,
-    });
-  }
-
-  onModalChatUsersClose(): void {
-    this.setProps({
-      modalChatUsersIsOpened: false,
-    });
-  }
-
-  onModalAddUserOpen(): void {
-    this.setProps({
-      modalAddUserIsOpened: true,
-      popupIsOpened: false,
-    });
-  }
-
-  onModalAddUserClose(): void {
-    this.setProps({
-      modalAddUserIsOpened: false,
-    });
-  }
-
   deleteChats(): void {
     if (!this.props.activeChat?.id) {
       return;
@@ -138,60 +81,18 @@ export class ChatDetailsComponent extends Block<ChatDetailsComponentProps> {
     });
   }
 
-  deleteUser(event: MouseEvent): void {
-    const target = event.target as HTMLElement;
-    const parent = target?.parentNode as HTMLDivElement;
-    if (!parent) {
-      return;
-    }
-
-    const userId = parseInt(parent.dataset['userId'] ?? '0', 10);
-    if (!userId || Number.isNaN(userId)) {
-      return;
-    }
-
-    this.props.store.dispatch(ChatsService.removeUser, {
-      users: [userId],
-      chatId: this.props.activeChat?.id,
-    });
+  // Add user
+  onModalAddUserOpen(): void {
+    this.onPopupClose();
+    const addUserModal = this.refs['addUserModal'] as ModalAddUserComponent;
+    addUserModal.onModalAddUserOpen();
   }
 
-  addUser(event: MouseEvent): void {
-    const target = event.target as HTMLElement;
-    const parent = target?.parentNode as HTMLDivElement;
-    if (!parent) {
-      return;
-    }
-
-    const userId = parseInt(parent.dataset['userId'] ?? '0', 10);
-    if (!userId || Number.isNaN(userId)) {
-      return;
-    }
-
-    console.log('Add User -> ', userId);
-    this.props.store.dispatch(ChatsService.addUser, {
-      users: [userId],
-      chatId: this.props.activeChat?.id,
-    });
-  }
-
-  onInputSearchUser(event: InputEvent): void {
-    console.log('InputEvent -> ', event);
-    const target = event.target as HTMLInputElement;
-    if (!target) {
-      return;
-    }
-
-    this.inputSearchUser = target.value;
-  }
-
-  onSearchUser(event: MouseEvent): void {
-    event?.preventDefault();
-    // searchUser
-    this.props.store.dispatch(UserService.searchUser, {
-      login: this.inputSearchUser,
-    });
-    console.log('MouseEvent -> ', event);
+  // Chat users
+  onModalChatUsersOpen(): void {
+    this.onPopupClose();
+    const chatUsersModal = this.refs['chatUsersModal'] as ModalChatUsersComponent;
+    chatUsersModal.onModalChatUsersOpen();
   }
 
   protected override render(): string {
@@ -255,62 +156,10 @@ export class ChatDetailsComponent extends Block<ChatDetailsComponentProps> {
             </div>
 
             <!--Modals chat users-->
-            {{#ModalComponent ref='chatUsersModal'
-                              isOpened=modalChatUsersIsOpened
-                              onClose=onModalChatUsersClose}}
-                <div class='list-users'>
-                    {{#each chatUsers}}
-                        <div class='list-users__user' data-user-id='{{this.id}}'>
-                            <div class='list-users__avatar'>
-                                {{{UserAvatarComponent avatar=this.avatar}}}
-                            </div>
-                            <span>{{this.firstName}} {{this.secondName}}</span>
-
-                            {{{ButtonComponent title='Удалить'
-                                               onClick=../deleteUser}}}
-                        </div>
-                    {{/each}}
-                </div>
-            {{/ModalComponent}}
+            {{{ModalChatUsersComponent ref='chatUsersModal'}}}
 
             <!--Modals add user-->
-            {{#ModalComponent ref='addUserModal'
-                              isOpened=modalAddUserIsOpened
-                              onClose=onModalAddUserClose}}
-
-                <form action='#' class='form'>
-                    {{{InputComponent
-                            label='Логин'
-                            className='form__input'
-                            id='addUserInput'
-                            type='text'
-                            name='login'
-                            placeholder=''
-                            onInput=onInputSearchUser
-                            validate=validateRuleType.Message
-                    }}}
-
-                    {{{ButtonComponent type='submit'
-                                       title='Найти'
-                                       onClick=onSearchUser}}}
-
-
-                    <div class='list-users'>
-                        {{#each searchUsers}}
-                            <div class='list-users__user' data-user-id='{{this.id}}'>
-                                <div class='list-users__avatar'>
-                                    {{{UserAvatarComponent avatar=this.avatar}}}
-                                </div>
-                                <span>{{this.firstName}} {{this.secondName}}</span>
-
-                                {{{ButtonComponent title='Добавить'
-                                                   onClick=../addUser}}}
-                            </div>
-                        {{/each}}
-                    </div>
-                </form>
-
-            {{/ModalComponent}}
+            {{{ModalAddUserComponent ref='addUserModal'}}}
 
         </div>
     `;
@@ -321,9 +170,6 @@ export default routerHoc(
   activeChatHoc(
     storeHoc(ChatDetailsComponent, (state) => ({
       activeChat: state.activeChat,
-      chatUsers: state.chatUsers,
-      searchUsers: state.searchUsers,
-      searchUsersFormError: state.searchUsersFormError,
     })),
   ),
 );
