+++ src/components/modal-add-user/index.ts
@@ -1,0 +1,1 @@
+export { default as ModalAddUserComponent } from './modal-add-user.component';
