+++ src/components/modal-chat-users/index.ts
@@ -1,0 +1,1 @@
+export { default as ModalChatUsersComponent } from './modal-chat-users.component';
